Start networkAssetTools, refactor asset validation files
