+++ src/ducks/modules/protocol/assetManifest.js
@@ -4,7 +4,7 @@ import path from 'path';
 import log from 'electron-log';
 import { importAsset as fsImportAsset } from '@app/other/protocols';
 import { getActiveProtocolMeta } from '@selectors/protocols';
-import { validateAsset } from '@app/other/protocols/importAsset';
+import validateAsset from '@app/other/protocols/validateAsset';
 import { invalidAssetErrorDialog, importAssetErrorDialog } from '@modules/protocol/utils/dialogs';
 
 const IMPORT_ASSET = 'PROTOCOL/IMPORT_ASSET';
