+++ src/other/protocols/importAsset.js
@@ -1,10 +1,8 @@
 import path from 'path';
 import uuid from 'uuid/v1';
-import { findKey, get, toLower, first } from 'lodash';
-import csvParse from 'csv-parse';
-import { copy, readFile } from 'fs-extra';
+import { findKey, toLower } from 'lodash';
+import { copy } from 'fs-extra';
 import { SUPPORTED_EXTENSION_TYPE_MAP } from '@app/config';
-import { getVariableNamesFromNetwork, validateNames } from '@app/protocol-validation/validation/validateExternalData';
 
 /**
  * Function that determines the type of an asset file when importing. Types are defined
@@ -16,7 +14,7 @@ import { getVariableNamesFromNetwork, validateNames } from '@app/protocol-valida
  * @return {string} Returns one of network, image, audio, video or returns false if type
  * is unsupported
  */
-const getSupportedAssetType = (filePath) => {
+export const getSupportedAssetType = (filePath) => {
   const extension = toLower(path.extname(filePath));
 
   const typeFromMap = findKey(SUPPORTED_EXTENSION_TYPE_MAP, type => type.includes(extension));
@@ -39,89 +37,5 @@ const importAsset = (protocolPath, filePath) =>
       .then(() => ({ filePath: destinationName, assetType }))
       .then(resolve);
   });
-
-const validateJson = data =>
-  new Promise((resolve, reject) => {
-    try {
-      const network = JSON.parse(data);
-
-      if (get(network, 'nodes', []).length === 0 && get(network, 'edges', []).length === 0) {
-        return reject(new Error("JSON network asset doesn't include any nodes or edges"));
-      }
-
-      // check variable names
-      const variableNames = getVariableNamesFromNetwork(network);
-
-      const error = validateNames(variableNames);
-
-      if (error) { return reject(new Error(error)); }
-
-      // Check option values
-      // Not needed at this point, as unsupported feature
-
-      return resolve(true);
-    } catch (e) {
-      return reject(e);
-    }
-  });
-
-const validateCsv = data =>
-  new Promise((resolve, reject) => {
-    try {
-      csvParse(data, (err, tableData) => {
-        if (err) {
-          return reject(new Error(err));
-        }
-
-        const firstRow = first(tableData);
-
-        // Check heading (variables) are valid
-        const error = validateNames(firstRow);
-
-        if (error) { return reject(new Error(error)); }
-
-        return resolve(true);
-      });
-    } catch (e) {
-      reject(e);
-    }
-  });
-
-/**
- * Get validator based on filePath, if no validator available it resolves by default
- * @param {string} filepath - The filename of the network asset
- * @returns {string} - Returns a function that returns a promise.
- */
-const getAssetValidator = (filePath) => {
-  const extension = path.extname(filePath).substr(1); // e.g. 'csv'
-
-  switch (extension) {
-    case 'csv':
-      return validateCsv;
-    case 'json':
-      return validateJson;
-    default:
-      // Don't validate non-network assets at this time
-      return () => Promise.resolve();
-  }
-};
-
-/**
- * Checks that imported asset is valid
- * @param {buffer} file - The file to check.
- */
-export const validateAsset = (filePath) => {
-  const assetType = getSupportedAssetType(filePath);
-
-  // Handle unsupported filePath types
-  if (!assetType) {
-    return Promise.reject(new Error('Asset type not supported'));
-  }
-
-  const assetValidator = getAssetValidator(filePath);
-
-  return readFile(filePath)
-    .then(assetValidator);
-};
 
 export default importAsset;
