+++ src/other/protocols/networkAssetTools.js
@@ -1,0 +1,75 @@
+/* eslint-disable import/prefer-default-export */
+
+import path from 'path';
+import { first, get } from 'lodash';
+import csvParse from 'csv-parse';
+import { readFile } from 'fs-extra';
+import { getVariableNamesFromNetwork } from '@app/protocol-validation/validation/validateExternalData';
+import { getSupportedAssetType } from '@app/other/protocols/importAsset';
+
+
+/**
+* Generate a switching function that takes a filepath as an argument
+* and returns match from configuration object.
+*/
+const withExtensionSwitch = (configuration, fallback = () => Promise.resolve()) =>
+  (filePath) => {
+    const extension = path.extname(filePath).substr(1); // e.g. 'csv'
+
+    return get(configuration, [extension], fallback);
+  };
+
+
+const readJsonVariables = data =>
+  new Promise((resolve, reject) => {
+    try {
+      const network = JSON.parse(data);
+
+      const variableNames = getVariableNamesFromNetwork(network);
+
+      return resolve(variableNames);
+    } catch (e) {
+      return reject(e);
+    }
+  });
+
+const readCsvVariables = data =>
+  new Promise((resolve, reject) => {
+    try {
+      csvParse(data, { trim: true }, (err, tableData) => {
+        const firstRow = first(tableData);
+
+        return resolve(firstRow);
+      });
+    } catch (e) {
+      reject(e);
+    }
+  });
+
+/**
+* Get validator based on filePath, if no validator available it resolves by default
+* @param {string} filepath - The filename of the network asset
+* @returns {string} - Returns a function that returns a promise.
+*/
+const getVariableReader = withExtensionSwitch({
+  csv: readCsvVariables,
+  json: readJsonVariables,
+});
+
+/**
+* Checks that imported asset is valid
+* @param {buffer} file - The file to check.
+*/
+export const getVariables = (filePath) => {
+  const assetType = getSupportedAssetType(filePath);
+
+  // Handle unsupported filePath types
+  if (!assetType) {
+    return Promise.reject(new Error('Asset type not supported'));
+  }
+
+  const variableReader = getVariableReader(filePath);
+
+  return readFile(filePath)
+    .then(variableReader);
+};
