+++ src/other/protocols/pruneAssets.js
